+++ app/Main.hs
@@ -10,7 +10,7 @@ import Control.Exception (throwIO)
 import Control.Monad
 import Data.Bool (bool)
 import Data.List (intercalate, sort)
-import Data.Maybe (mapMaybe)
+import Data.Maybe (fromMaybe, mapMaybe)
 import qualified Data.Text as T
 import qualified Data.Text.IO as TIO
 import Data.Version (showVersion)
@@ -32,7 +32,12 @@ import System.IO (hPutStrLn, stderr)
 main :: IO ()
 main = do
   Opts {..} <- execParser optsParserInfo
-  let formatOne' = formatOne optCabalDefaultExtensions optMode optConfig
+  let formatOne' =
+        formatOne
+          optCabalDefaultExtensions
+          optMode
+          optSourceType
+          optConfig
   exitCode <- case optInputFiles of
     [] -> formatOne' Nothing
     ["-"] -> formatOne' Nothing
@@ -59,18 +64,20 @@ formatOne ::
   CabalDefaultExtensionsOpts ->
   -- | Mode of operation
   Mode ->
+  -- | The 'SourceType' requested by the user
+  Maybe SourceType ->
   -- | Configuration
   Config RegionIndices ->
   -- | File to format or stdin as 'Nothing'
   Maybe FilePath ->
   IO ExitCode
-formatOne CabalDefaultExtensionsOpts {..} mode config mpath =
-  withPrettyOrmoluExceptions (cfgColorMode config) $
+formatOne CabalDefaultExtensionsOpts {..} mode reqSourceType rawConfig mpath =
+  withPrettyOrmoluExceptions (cfgColorMode rawConfig) $ do
     case FP.normalise <$> mpath of
       -- input source = STDIN
       Nothing -> do
         resultConfig <-
-          configPlus
+          patchConfig Nothing
             <$> if optUseCabalDefaultExtensions
               then case optStdinInputFile of
                 Just stdinInputFile ->
@@ -97,7 +104,7 @@ formatOne CabalDefaultExtensionsOpts {..} mode config mpath =
       -- input source = a file
       Just inputFile -> do
         resultConfig <-
-          configPlus
+          patchConfig (Just (detectSourceType inputFile))
             <$> if optUseCabalDefaultExtensions
               then getCabalExtensionDynOptions inputFile
               else pure []
@@ -108,22 +115,31 @@ formatOne CabalDefaultExtensionsOpts {..} mode config mpath =
           InPlace -> do
             -- ormoluFile is not used because we need originalInput
             originalInput <- readFileUtf8 inputFile
-            formattedInput <- ormolu resultConfig inputFile (T.unpack originalInput)
+            formattedInput <-
+              ormolu resultConfig inputFile (T.unpack originalInput)
             when (formattedInput /= originalInput) $
               writeFileUtf8 inputFile formattedInput
             return ExitSuccess
           Check -> do
             -- ormoluFile is not used because we need originalInput
             originalInput <- readFileUtf8 inputFile
-            formattedInput <- ormolu resultConfig inputFile (T.unpack originalInput)
+            formattedInput <-
+              ormolu resultConfig inputFile (T.unpack originalInput)
             handleDiff originalInput formattedInput inputFile
   where
-    configPlus dynOpts = config {cfgDynOptions = cfgDynOptions config ++ dynOpts}
+    patchConfig mdetectedSourceType dynOpts =
+      rawConfig
+        { cfgDynOptions = cfgDynOptions rawConfig ++ dynOpts,
+          cfgSourceType =
+            fromMaybe
+              ModuleSource
+              (reqSourceType <|> mdetectedSourceType)
+        }
     handleDiff originalInput formattedInput fileRepr =
       case diffText originalInput formattedInput fileRepr of
         Nothing -> return ExitSuccess
         Just diff -> do
-          runTerm (printTextDiff diff) (cfgColorMode config) stderr
+          runTerm (printTextDiff diff) (cfgColorMode rawConfig) stderr
           -- 100 is different to all the other exit code that are emitted
           -- either from an 'OrmoluException' or from 'error' and
           -- 'notImplemented'.
@@ -139,6 +155,8 @@ data Opts = Opts
     optConfig :: !(Config RegionIndices),
     -- | Options for respecting default-extensions from .cabal files
     optCabalDefaultExtensions :: CabalDefaultExtensionsOpts,
+    -- | Source type option, where 'Nothing' means autodetection
+    optSourceType :: !(Maybe SourceType),
     -- | Haskell source files to format or stdin (when the list is empty)
     optInputFiles :: ![FilePath]
   }
@@ -213,6 +231,7 @@ optsParser =
         )
     <*> configParser
     <*> cabalDefaultExtensionsParser
+    <*> sourceTypeParser
     <*> (many . strArgument . mconcat)
       [ metavar "FILE",
         help "Haskell source files to format or stdin (the default)"
@@ -255,6 +274,10 @@ configParser =
         short 'c',
         help "Fail if formatting is not idempotent"
       ]
+    -- We cannot parse the source type here, because we might need to do
+    -- autodection based on the input file extension (not available here)
+    -- before storing the resolved value in the config struct.
+    <*> pure ModuleSource
     <*> (option parseColorMode . mconcat)
       [ long "color",
         metavar "WHEN",
@@ -274,6 +297,16 @@ configParser =
               ]
         )
 
+sourceTypeParser :: Parser (Maybe SourceType)
+sourceTypeParser =
+  (option parseSourceType . mconcat)
+    [ long "source-type",
+      short 't',
+      metavar "TYPE",
+      value Nothing,
+      help "Set the type of source; TYPE can be 'module', 'sig', or 'auto' (the default)"
+    ]
+
 ----------------------------------------------------------------------------
 -- Helpers
 
@@ -285,9 +318,19 @@ parseMode = eitherReader $ \case
   "check" -> Right Check
   s -> Left $ "unknown mode: " ++ s
 
+-- | Parse 'ColorMode'.
 parseColorMode :: ReadM ColorMode
 parseColorMode = eitherReader $ \case
   "never" -> Right Never
   "always" -> Right Always
   "auto" -> Right Auto
   s -> Left $ "unknown color mode: " ++ s
+
+-- | Parse the 'SourceType'. 'Nothing' means that autodetection based on
+-- file extension is requested.
+parseSourceType :: ReadM (Maybe SourceType)
+parseSourceType = eitherReader $ \case
+  "module" -> Right (Just ModuleSource)
+  "sig" -> Right (Just SignatureSource)
+  "auto" -> Right Nothing
+  s -> Left $ "unknown source type: " ++ s